Switched to fragment based search UI
