+++ app/src/main/res/layout/fragment_home.xml
@@ -7,14 +7,14 @@
     tools:context=".ui.main.view.MainActivity">
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recViewFeaturedCategories"
+        android:id="@+id/recyclerViewFeaturedCategories"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <ProgressBar
-        android:id="@+id/progBarFeaturedCategories"
+        android:id="@+id/progressBarFeaturedCategories"
         style="?android:attr/progressBarStyle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
