+++ app/src/main/res/navigation/search_navigation.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/search_nav_graph"
+    app:startDestination="@id/navigation_results">
+
+    <fragment
+        android:id="@+id/navigation_results"
+        android:name="com.sthoray.allright.ui.search.view.ResultsFragment"
+        android:label="@string/title_results"
+        tools:layout="@layout/fragment_home" />
+
+</navigation>
