+++ app/src/main/res/values/strings.xml
@@ -6,8 +6,11 @@
     <!-- Format strings-->
     <string name="format_price">$%.2f</string>
 
-    <!-- Menu items -->
+    <!-- Main menu items -->
     <string name="title_home">Home</string>
     <string name="title_browse">Browse</string>
     <string name="title_profile">Profile</string>
+
+    <!-- Search fragments -->
+    <string name="title_results">Search Results</string>
 </resources>
